Move test utils out of model test into common.py
